adding Swal in delete edit and add todo

+++ src/components/SingleTodo.vue
@@ -27,6 +27,7 @@
 <script>
 import { mapActions } from "vuex";
 import { ref } from "@vue/reactivity";
+import Swal from "sweetalert2";
 export default {
   props: ["todo"],
   setup() {
@@ -41,7 +42,20 @@ export default {
       this.doneBtn(todo);
     },
     async deleteTodo(id) {
-      await this.deleteTodoItem(id);
+      await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.deleteTodoItem(id);
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        }
+      });
     },
     ...mapActions(["doneBtn", "deleteTodoItem"]),
   },
